ml-server/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
yfinance==0.2.18
requests==2.31.0
python-dotenv==1.0.0
tensorflow==2.15.0
keras==2.15.0
transformers==4.35.0
torch==2.1.0
textblob==0.17.1
vaderSentiment==3.3.2